+++ lib/infinity_widgets.dart
@@ -22,6 +22,7 @@ export 'src/widgets/lazy_indexed_stack.dart';
 export 'src/widgets/list_item.dart';
 export 'src/widgets/list_item_chevron.dart';
 export 'src/widgets/modals/dialog_header.dart';
+export 'src/widgets/modals/message_dialog.dart';
 export 'src/widgets/modals/modal_sheet.dart';
 export 'src/widgets/modals/modal_sheet_page.dart';
 export 'src/widgets/responsive/responsive_scaffold.dart';