+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Added color argument to getForegroundColor
 - Added filled constructor to Button
+- Added message dialog
 
 ## 1.0.0